requirements.txt
opencv_python==4.2.0.34
numpy==1.19.5
tensorflow==2.1.0
scikit-video==1.1.11
tf_slim==1.1.0
ffmpeg-python==0.2.0